Plantify new/plantify/requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
Werkzeug==3.0.1 